Fix Ruby 1.9 / Rails 4.x Travis CI build
Recently mime-types 3 included a dependency on mime-types-data which is Ruby 2+
only. When bundler resolved the bundle update on travis it picked the latest
version of mime-types even when being run on Ruby 1.9 versions. This should
hopefully resolve that issue by requiring a version of mime types <= 3.0 for
the combination of Ruby 1.9 and Rails 4.x.