Adds Gemfile for Rails 6 and generic stable rails
With the recent release of Rails 6 we want to make sure everything works
as expected. I also plan to remove the upper bound because the Rails API
is pretty stable and less likely to break. Adding the rails-stable
gemfile and enabling periodic runs on Travis should be enough to notify
us when things actually break.